% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clamp}
\alias{clamp}
\title{Clamp a vector between some modulus}
\usage{
clamp(x, modulus = 5)
}
\arguments{
\item{x}{numeric vector}

\item{modulus}{numeric scalar}
}
\value{
numeric vector
}
\description{
Replaces all values greater than modulus with modulus and all values lt modulus with -modulus.
Also known as winsorization.
}
