% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrices.R
\name{fisher_corr}
\alias{fisher_corr}
\alias{print.AMcorr}
\title{Generate a correlation matrix and confidence intervals}
\usage{
fisher_corr(x, y, conf.level = 0.95, ..., method = "fisher")

\method{print}{AMcorr}(
  x,
  cor_fmt = "\%.2f",
  sep = " ",
  include_ci = FALSE,
  include_pstar = FALSE,
  pstar_levels = c(`*` = 0.05, `**` = 0.01),
  ...
)
}
\arguments{
\item{x}{\code{AMcorr}}

\item{y}{numeric matrix-like object. If missing \code{y=x}.}

\item{conf.level}{numeric confidence level}

\item{...}{ignored}

\item{method}{inference method, currently only 'fisher' is supported}

\item{cor_fmt}{an sprintf format string for correlation}

\item{sep}{separators between correlation and confidence intervals, and/or pvalues}

\item{include_ci}{include upper/lower CI?}

\item{include_pstar}{include pvalue asterix (set according to \code{pstar_levels})}

\item{pstar_levels}{named numeric vector determining pvalue formatting}
}
\value{
structure containing correlation matrix, \code{lr}, \code{ur} lower and upper CI, the pvalues from the associated tests \code{pz} and the degrees of freedom \code{dof}.

character matrix
}
\description{
Treatment of NAs is incomplete at this time (in particular, the degrees of freedom will not be correct).
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Pretty print a correlation matrix and optionally its confidence intervals and p values
}}

\examples{
data(iris)
fc = fisher_corr(iris[, c('Sepal.Length', 'Sepal.Width')],
iris[, c('Petal.Length', 'Petal.Width')])
print(fc, include_pstar = TRUE)
print(fc, include_pstar = TRUE, include_ci = TRUE, cor_fmt = '\%.1f')
}
