% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_methods.R
\name{gsea_by_group}
\alias{gsea_by_group}
\alias{as.data.frame.GroupedGSEA}
\alias{plot.GroupedGSEA}
\title{Run a GSEA by group}
\usage{
gsea_by_group(
  formula,
  gene_identifier,
  data,
  set_list,
  nperm = 500,
  maxSize = 500,
  ...
)

\method{as.data.frame}{GroupedGSEA}(
  x,
  row.names = NULL,
  optional = TRUE,
  ...,
  topn = Inf,
  wrap_len = 30,
  p.adjust_method = "BH"
)

\method{plot}{GroupedGSEA}(x, ..., topn = 5, wrap_len = 30, p.adjust_method = "BH")
}
\arguments{
\item{formula}{a \code{formula} of the kind \code{numeric_score ~ group1 + group2 + ...} identifying the column containing the score vs the grouping variables, which will be interacted with each other}

\item{gene_identifier}{\code{character}}

\item{data}{\code{data.frame} in which the \code{formula} and \code{gene_identifer} are interpreted}

\item{set_list}{alternatively, a list of named and sorted input}

\item{nperm}{Number of permutations to do. Minimial possible nominal p-value is about 1/nperm}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{...}{passed to \code{\link[fgsea:fgsea]{fgsea::fgsea()}}}

\item{x}{object of class \code{GroupedGSEA}}

\item{row.names}{ignored}

\item{optional}{ignored}

\item{topn}{report at least this many pathways (by minimum p value across groups)}

\item{wrap_len}{wrap pathway titles after this many characters}

\item{p.adjust_method}{passed to \code{\link[stats:p.adjust]{stats::p.adjust()}}}
}
\value{
an object of class \code{GroupedGSEA}, which is internally just a list \code{\link[fgsea:fgsea]{fgsea::fgsea()}} objects.
}
\description{
Run a GSEA by group
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.data.frame}: coerce result to a \code{data.frame}

\item \code{plot}: make a plot of the results
}}

\examples{
data(exampleRanks, package = 'fgsea')
data(examplePathways, package = 'fgsea')
data = data.frame(score = exampleRanks, gene = names(exampleRanks), treatment = rep(c('A', 'B'), length.out = length(exampleRanks)))
gout = gsea_by_group(score ~ treatment, 'gene', data, pathways = examplePathways)
df = as.data.frame(gout)
plot(gout)
}
