% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_control.R
\name{h2}
\alias{h2}
\title{Examine the upper left corner of a matrix-like object}
\usage{
h2(x, p = 15)
}
\arguments{
\item{x}{a matrix or data.frame}

\item{p}{number of rows/columns to examine}
}
\value{
a slice of x
}
\description{
Examine the upper left corner of a matrix-like object
}
