% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{heat2}
\alias{heat2}
\title{A quick and dirty heatmap}
\usage{
heat2(
  mat,
  Rowv = NA,
  Colv = NA,
  symbreaks = TRUE,
  col = if (symbreaks) gplots::redblue else gplots::redgreen,
  ...
)
}
\arguments{
\item{mat}{numeric matrix}

\item{Rowv}{determines if and how the row dendrogram should be reordered.    By default, it is TRUE, which implies dendrogram is computed and reordered based on row means. If NULL or FALSE, then no dendrogram is computed and no reordering is done. If a dendrogram, then it is used "as-is", ie without any reordering. If a vector of integers, then dendrogram is computed and reordered based on the order of the vector.}

\item{Colv}{determines if and how the column dendrogram should be reordered.    Has the options as the Rowv argument above and additionally when x is a square matrix, Colv="Rowv" means that columns should be treated identically to the rows.}

\item{symbreaks}{Boolean indicating whether breaks should be made symmetric about 0. Defaults to TRUE if the data includes negative values, and to FALSE otherwise.}

\item{col}{colors used for the image.}

\item{...}{passed to gplots::heatmap.2}
}
\value{
make a plot
}
\description{
A quick and dirty heatmap
}
