% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{purify}
\alias{purify}
\title{Largest elements in norm from a matrix}
\usage{
purify(x, top = floor(0.1 * nrow(x)), remove_zero = FALSE, p = 2)
}
\arguments{
\item{x}{numeric matrix}

\item{top}{number of top elements}

\item{remove_zero}{\code{logical}}

\item{p}{Coefficient \code{p} of the L_p norm}
}
\description{
The \code{top} largest rows in norm from \code{x} are retained.
Others are set to zero or removed (\code{remove_zero})
}
