% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_functions.R
\name{rrndr}
\alias{rrndr}
\title{Shortcut for rmarkdown::render that preserves intermediates}
\usage{
rrndr(..., envir = parent.frame(), clean = FALSE)
}
\arguments{
\item{...}{passed to rmarkdown}

\item{envir}{The environment in which the code chunks are to be evaluated during knitting (can use new.env() to guarantee an empty new environment).}

\item{clean}{TRUE to clean intermediate files created during rendering.}
}
\value{
When run_pandoc = TRUE, the compiled document is written into the output file, and the path of the output file is returned.
When run_pandoc = FALSE, the path of the Markdown output file, with attributes knit_meta (the knitr meta data collected from code chunks) and intermediates (the intermediate files/directories generated by render()).
}
\description{
Shortcut for rmarkdown::render that preserves intermediates
}
