% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_aesthetics.R
\name{set_AM_heatmap_defaults}
\alias{set_AM_heatmap_defaults}
\title{Override the default color mappings in ComplexHeatmap
This function naughtily re-assigns a modified version of default_col into the ComplexHeatmap namespace
It will probably break in subsequent versions of complexheatmap.}
\usage{
set_AM_heatmap_defaults()
}
\value{
none
}
\description{
Override the default color mappings in ComplexHeatmap
This function naughtily re-assigns a modified version of default_col into the ComplexHeatmap namespace
It will probably break in subsequent versions of complexheatmap.
}
